myadk/requirements.txt
django 
django-jazzmin
djangorestframework 
google-adk 
gunicorn
